Adding server script to NetworkManager
This was done in the server project, I have set the port to 7777 and max client count to 10. You can adjust port according to the list in the "Multiplayer files" folder where a txt contains a link to unassigned ones. When you start the game a message appears in the console saying you have started a server using the specified port.